+++ sentry/sentry.py
@@ -1,10 +1,17 @@
 import sentry_sdk
 
 sentry_sdk.init(
-  dsn="https://3f05934c49e0a8f851715c05a7c34b31@o4505677157695488.ingest.sentry.io/4505681394073600",
-
-  # Set traces_sample_rate to 1.0 to capture 100%
-  # of transactions for performance monitoring.
-  # We recommend adjusting this value in production.
-  traces_sample_rate=1.0
+    dsn="https://3f05934c49e0a8f851715c05a7c34b31@o4505677157695488.ingest.sentry.io/4505681394073600",
+    traces_sample_rate=1.0
 )
+
+def zero_division():
+    try:
+        zerocheck= int(input("Enter the number: "))
+        zerocheck = 1 / zerocheck
+        print(zerocheck)
+    except Exception as e:
+        sentry_sdk.capture_exception(e)
+        sentry_sdk.capture_message("Zero Division Error en la otra pagina")
+
+zero_division()
